compile with more warnings and change some types to size_t instead of int. Compiles but could lead to wrong behaviour maybe

+++ examples/cifar.c
@@ -14,7 +14,7 @@ void train_cifar(char *cfgfile, char *weightfile)
     int N = 50000;
 
     char **labels = get_labels("data/cifar/labels.txt");
-    int epoch = (*net->seen)/N;
+    size_t epoch = (*net->seen)/N;
     data train = load_all_cifar10();
     while(get_current_batch(net) < net->max_batches || net->max_batches == 0){
         clock_t time=clock();
@@ -26,7 +26,7 @@ void train_cifar(char *cfgfile, char *weightfile)
         if(*net->seen/N > epoch){
             epoch = *net->seen/N;
             char buff[256];
-            sprintf(buff, "%s/%s_%d.weights",backup_directory,base, epoch);
+            sprintf(buff, "%s/%s_%ld.weights",backup_directory,base, epoch);
             save_weights(net, buff);
         }
         if(get_current_batch(net)%100 == 0){
@@ -59,7 +59,7 @@ void train_cifar_distill(char *cfgfile, char *weightfile)
     int N = 50000;
 
     char **labels = get_labels("data/cifar/labels.txt");
-    int epoch = (*net->seen)/N;
+    size_t epoch = (*net->seen)/N;
 
     data train = load_all_cifar10();
     matrix soft = csv_to_matrix("results/ensemble.csv");
@@ -79,7 +79,7 @@ void train_cifar_distill(char *cfgfile, char *weightfile)
         if(*net->seen/N > epoch){
             epoch = *net->seen/N;
             char buff[256];
-            sprintf(buff, "%s/%s_%d.weights",backup_directory,base, epoch);
+            sprintf(buff, "%s/%s_%ld.weights",backup_directory,base, epoch);
             save_weights(net, buff);
         }
         if(get_current_batch(net)%100 == 0){
